Unify CC0016 and CC0031
Closes #662
