+++ src/Common/CodeCracker.Common/DiagnosticId.cs
@@ -18,7 +18,6 @@
         TernaryOperator_Return = 13,
         TernaryOperator_Assignment = 14,
         UnnecessaryParenthesis = 15,
-        CopyEventToVariableBeforeFire = 16,
         SwitchToAutoProp = 17,
         ExistenceOperator = 18,
         ConvertToSwitch = 19,
