+++ test/CSharp/CodeCracker.Test/CodeCracker.Test.csproj
@@ -162,7 +162,6 @@
     <Compile Include="Usage\AbstractClassShouldNotHavePublicCtorTests.cs" />
     <Compile Include="Usage\ArgumentExceptionTests.cs" />
     <Compile Include="Style\ConvertToSwitchTests.cs" />
-    <Compile Include="Design\CopyEventToVariableBeforeFireTests.cs" />
     <Compile Include="Usage\CallExtensionMethodAsExtensionTests.cs" />
     <Compile Include="Usage\DisposableVariableNotDisposedTests.cs" />
     <Compile Include="Usage\DisposablesShouldCallSuppressFinalizeTests.cs" />
