+++ test/CSharp/CodeCracker.Test/Design/UseInvokeMethodToFireEventTests.cs
@@ -57,6 +57,350 @@ namespace CodeCracker.Test.CSharp.Design
         }
 
         [Fact]
+        public async void WarningIfEventIsReadOnlyAndAssignedInIfInConstructor()
+        {
+            const string test = @"
+                public class MyClass
+                {
+                    readonly System.EventHandler MyEvent;
+
+                    public MyClass(bool shouldAssign)
+                    {
+                        if(shouldAssign)
+                        {
+                            MyEvent = (sender, args) => { };
+                        }
+                    }
+
+                    public void Execute()
+                    {
+                        MyEvent(this, System.EventArgs.Empty);
+                    }
+                }";
+
+            var expected = new DiagnosticResult
+            {
+                Id = DiagnosticId.UseInvokeMethodToFireEvent.ToDiagnosticId(),
+                Message = string.Format(UseInvokeMethodToFireEventAnalyzer.MessageFormat.ToString(), "MyEvent"),
+                Severity = DiagnosticSeverity.Warning,
+                Locations = new[] { new DiagnosticResultLocation("Test0.cs", 16, 25) }
+            };
+
+            await VerifyCSharpDiagnosticAsync(test, expected);
+        }
+
+        [Fact]
+        public async void WarningIfEventIsReadOnlyAndAssignedInForeachInConstructor()
+        {
+            const string test = @"
+                public class MyClass
+                {
+                    readonly System.EventHandler MyEvent;
+
+                    public MyClass(int[] values)
+                    {
+                        foreach(var value in values)
+                        {
+                            MyEvent = (sender, args) => { };
+                        }
+                    }
+
+                    public void Execute()
+                    {
+                        MyEvent(this, System.EventArgs.Empty);
+                    }
+                }";
+
+            var expected = new DiagnosticResult
+            {
+                Id = DiagnosticId.UseInvokeMethodToFireEvent.ToDiagnosticId(),
+                Message = string.Format(UseInvokeMethodToFireEventAnalyzer.MessageFormat.ToString(), "MyEvent"),
+                Severity = DiagnosticSeverity.Warning,
+                Locations = new[] { new DiagnosticResultLocation("Test0.cs", 16, 25) }
+            };
+
+            await VerifyCSharpDiagnosticAsync(test, expected);
+        }
+
+        [Fact]
+        public async void WarningIfEventIsReadOnlyAndAssignedInForInConstructor()
+        {
+            const string test = @"
+                public class MyClass
+                {
+                    readonly System.EventHandler MyEvent;
+
+                    public MyClass(int number)
+                    {
+                        for(int i = 0; i < number; i++)
+                        {
+                            MyEvent = (sender, args) => { };
+                        }
+                    }
+
+                    public void Execute()
+                    {
+                        MyEvent(this, System.EventArgs.Empty);
+                    }
+                }";
+
+            var expected = new DiagnosticResult
+            {
+                Id = DiagnosticId.UseInvokeMethodToFireEvent.ToDiagnosticId(),
+                Message = string.Format(UseInvokeMethodToFireEventAnalyzer.MessageFormat.ToString(), "MyEvent"),
+                Severity = DiagnosticSeverity.Warning,
+                Locations = new[] { new DiagnosticResultLocation("Test0.cs", 16, 25) }
+            };
+
+            await VerifyCSharpDiagnosticAsync(test, expected);
+        }
+
+        [Fact]
+        public async void WarningIfEventIsReadOnlyAndAssignedInWhileInConstructor()
+        {
+            const string test = @"
+                public class MyClass
+                {
+                    readonly System.EventHandler MyEvent;
+
+                    public MyClass(int number)
+                    {
+                        while(number > 0)
+                        {
+                            MyEvent = (sender, args) => { };
+                            number--;
+                        }
+                    }
+
+                    public void Execute()
+                    {
+                        MyEvent(this, System.EventArgs.Empty);
+                    }
+                }";
+
+            var expected = new DiagnosticResult
+            {
+                Id = DiagnosticId.UseInvokeMethodToFireEvent.ToDiagnosticId(),
+                Message = string.Format(UseInvokeMethodToFireEventAnalyzer.MessageFormat.ToString(), "MyEvent"),
+                Severity = DiagnosticSeverity.Warning,
+                Locations = new[] { new DiagnosticResultLocation("Test0.cs", 17, 25) }
+            };
+
+            await VerifyCSharpDiagnosticAsync(test, expected);
+        }
+
+        [Fact]
+        public async void WarningIfEventIsReadOnlyAndAssignedAfterReturnInConstructor()
+        {
+            const string test = @"
+                public class MyClass
+                {
+                    readonly System.EventHandler MyEvent;
+
+                    public MyClass(bool returnEarly)
+                    {
+                        if(returnEarly)
+                        {
+                            return;
+                        }
+                        MyEvent = (sender, args) => { };
+                    }
+
+                    public void Execute()
+                    {
+                        MyEvent(this, System.EventArgs.Empty);
+                    }
+                }";
+
+            var expected = new DiagnosticResult
+            {
+                Id = DiagnosticId.UseInvokeMethodToFireEvent.ToDiagnosticId(),
+                Message = string.Format(UseInvokeMethodToFireEventAnalyzer.MessageFormat.ToString(), "MyEvent"),
+                Severity = DiagnosticSeverity.Warning,
+                Locations = new[] { new DiagnosticResultLocation("Test0.cs", 17, 25) }
+            };
+
+            await VerifyCSharpDiagnosticAsync(test, expected);
+        }
+
+        [Fact]
+        public async void WarningIfEventIsReadOnlyAndAssignedToNullRegularAssignmentOnFieldDeclaration()
+        {
+            const string test = @"
+                public class MyClass
+                {
+                    readonly System.EventHandler MyEvent = null;
+
+                    public void Execute()
+                    {
+                        MyEvent(this, System.EventArgs.Empty);
+                    }
+                }";
+
+            var expected = new DiagnosticResult
+            {
+                Id = DiagnosticId.UseInvokeMethodToFireEvent.ToDiagnosticId(),
+                Message = string.Format(UseInvokeMethodToFireEventAnalyzer.MessageFormat.ToString(), "MyEvent"),
+                Severity = DiagnosticSeverity.Warning,
+                Locations = new[] { new DiagnosticResultLocation("Test0.cs", 8, 25) }
+            };
+
+            await VerifyCSharpDiagnosticAsync(test, expected);
+        }
+
+        [Fact]
+        public async void WarningIfEventIsReadOnlyAndAssignedToNullInConstructor()
+        {
+            const string test = @"
+                public class MyClass
+                {
+                    readonly System.EventHandler MyEvent;
+
+                    public MyClass(bool returnEarly)
+                    {
+                        MyEvent = null;
+                    }
+
+                    public void Execute()
+                    {
+                        MyEvent(this, System.EventArgs.Empty);
+                    }
+                }";
+
+            var expected = new DiagnosticResult
+            {
+                Id = DiagnosticId.UseInvokeMethodToFireEvent.ToDiagnosticId(),
+                Message = string.Format(UseInvokeMethodToFireEventAnalyzer.MessageFormat.ToString(), "MyEvent"),
+                Severity = DiagnosticSeverity.Warning,
+                Locations = new[] { new DiagnosticResultLocation("Test0.cs", 13, 25) }
+            };
+
+            await VerifyCSharpDiagnosticAsync(test, expected);
+        }
+
+        [Fact]
+        public async void WarningIfEventIsReadOnlyAndAssignedToNullAfterRegularAssignmentInConstructor()
+        {
+            const string test = @"
+                public class MyClass
+                {
+                    readonly System.EventHandler MyEvent;
+
+                    public MyClass(bool returnEarly)
+                    {
+                        MyEvent = (sender, args) => { };
+                        MyEvent = null;
+                    }
+
+                    public void Execute()
+                    {
+                        MyEvent(this, System.EventArgs.Empty);
+                    }
+                }";
+
+            var expected = new DiagnosticResult
+            {
+                Id = DiagnosticId.UseInvokeMethodToFireEvent.ToDiagnosticId(),
+                Message = string.Format(UseInvokeMethodToFireEventAnalyzer.MessageFormat.ToString(), "MyEvent"),
+                Severity = DiagnosticSeverity.Warning,
+                Locations = new[] { new DiagnosticResultLocation("Test0.cs", 14, 25) }
+            };
+
+            await VerifyCSharpDiagnosticAsync(test, expected);
+        }
+
+        [Fact]
+        public async void WarningIfEventIsReadOnlyAndAssignedInAllSwitchCasesButNoDefaultInConstructor()
+        {
+            const string test = @"
+                public class MyClass
+                {
+                    readonly System.EventHandler MyEvent;
+
+                    public MyClass(int value)
+                    {
+                        switch(value)
+                        {
+                            case 1: MyEvent = (sender, args) => { }; break;
+                            case 2: MyEvent = (sender, args) => { }; break;
+                        }
+                    }
+
+                    public void Execute()
+                    {
+                        MyEvent(this, System.EventArgs.Empty);
+                    }
+                }";
+
+            var expected = new DiagnosticResult
+            {
+                Id = DiagnosticId.UseInvokeMethodToFireEvent.ToDiagnosticId(),
+                Message = string.Format(UseInvokeMethodToFireEventAnalyzer.MessageFormat.ToString(), "MyEvent"),
+                Severity = DiagnosticSeverity.Warning,
+                Locations = new[] { new DiagnosticResultLocation("Test0.cs", 17, 25) }
+            };
+
+            await VerifyCSharpDiagnosticAsync(test, expected);
+        }
+
+        [Fact]
+        public async void WarningIfEventIsReadOnlyAndAssignedInASwitchCaseButNotAllInConstructor()
+        {
+            const string test = @"
+                public class MyClass
+                {
+                    readonly System.EventHandler MyEvent;
+
+                    public MyClass(int value)
+                    {
+                        switch(value)
+                        {
+                            case 1: MyEvent = (sender, args) => { }; break;
+                            default: break;
+                        }
+                    }
+
+                    public void Execute()
+                    {
+                        MyEvent(this, System.EventArgs.Empty);
+                    }
+                }";
+
+            var expected = new DiagnosticResult
+            {
+                Id = DiagnosticId.UseInvokeMethodToFireEvent.ToDiagnosticId(),
+                Message = string.Format(UseInvokeMethodToFireEventAnalyzer.MessageFormat.ToString(), "MyEvent"),
+                Severity = DiagnosticSeverity.Warning,
+                Locations = new[] { new DiagnosticResultLocation("Test0.cs", 17, 25) }
+            };
+
+            await VerifyCSharpDiagnosticAsync(test, expected);
+        }
+
+        [Fact]
+        public async void NotWarningIfEventIsReadOnlyAndAssignedInConstructor()
+        {
+            const string test = @"
+                public class MyClass
+                {
+                    readonly System.EventHandler MyEvent;
+
+                    public MyClass()
+                    {
+                        MyEvent = (sender, args) => { };
+                    }
+
+                    public void Execute()
+                    {
+                        MyEvent(this, System.EventArgs.Empty);
+                    }
+                }";
+
+            await VerifyCSharpHasNoDiagnosticsAsync(test);
+        }
+
+
+        [Fact]
         public async void AcceptExpressionBodiedMethods()
         {
             const string test = @"
@@ -466,13 +810,289 @@ public static TReturn Method<T, TReturn>(System.Func<T, TReturn> getter) where T
         }
 
         [Fact]
-        public async void WhenMethodInvokedWithNonReferenceTypeHasNoDiagnostic()
+        public async void WhenMethodInvokedWithNonReferenceTypeHasOnlyIfNullDiagnostic()
         {
             var test = @"
+                public static TReturn Method<T, TReturn>(System.Func<T, TReturn> getter) where TReturn : struct
+                {
+                    return getter?.Invoke(default(T)) ?? default(TReturn);
+                }".WrapInCSharpClass();
+            await VerifyCSharpHasNumberOfCodeActions(test, 1);
+        }
+
+        public static TReturn Method<T, TReturn>(System.Func<T, TReturn> getter) where TReturn : struct
+        {
+            getter?.Invoke(default(T));
+            return getter?.Invoke(default(T)) ?? default(TReturn);
+        }
+
+        [Fact]
+        public async void FixWithInvokeWithNonReferenceType()
+        {
+            var source = @"
                 public static TReturn Method<T, TReturn>(System.Func<T, TReturn> getter) where T : System.Attribute where TReturn : struct
                 {
                     return getter(default(T));
                 }".WrapInCSharpClass();
+            var fix = @"
+                public static TReturn Method<T, TReturn>(System.Func<T, TReturn> getter) where T : System.Attribute where TReturn : struct
+                {
+                    return getter?.Invoke(default(T)) ?? default(TReturn);
+                }".WrapInCSharpClass();
+            await VerifyCSharpFixAsync(source, fix, codeFixIndex: 0);
+        }
+
+        [Fact]
+        public async void FixWithCheckForNullWithNonReferenceType()
+        {
+            var source = @"
+                public static TReturn Method<T, TReturn>(System.Func<T, TReturn> getter) where T : System.Attribute where TReturn : struct
+                {
+                    return getter(default(T));
+                }".WrapInCSharpClass();
+            var fix = @"
+                public static TReturn Method<T, TReturn>(System.Func<T, TReturn> getter) where T : System.Attribute where TReturn : struct
+                {
+                    var handler = getter;
+                    if (handler != null)
+                        return handler(default(T));
+                }".WrapInCSharpClass();
+            await VerifyCSharpFixAsync(source, fix, codeFixIndex: 1, allowNewCompilerDiagnostics: true);
+        }
+
+        [Fact]
+        public async void FixWithCheckForNullAndKeepCommentsWhenReplacedWithCodeFix()
+        {
+            const string source = @"
+                public class MyClass
+                {
+                    public event System.EventHandler MyEvent;
+
+                    public void Execute()
+                    {
+                        //comment
+                        MyEvent(this, System.EventArgs.Empty); //Some Comment
+                    }
+                }";
+
+            const string fixtest = @"
+                public class MyClass
+                {
+                    public event System.EventHandler MyEvent;
+
+                    public void Execute()
+                    {
+                        //comment
+                        var handler = MyEvent;
+                        if (handler != null)
+                            handler(this, System.EventArgs.Empty); //Some Comment
+                    }
+                }";
+            await VerifyCSharpFixAsync(source, fixtest, 1);
+        }
+
+        [Fact]
+        public async void FixWhenInvocationIsInsideABlockWithoutBraces()
+        {
+            const string source = @"
+                public class MyClass
+                {
+                    public event System.EventHandler MyEvent;
+                    bool raiseEvents = true;
+
+                    public void Execute()
+                    {
+                        if (raiseEvents) MyEvent(this, System.EventArgs.Empty);
+                    }
+                }";
+
+            const string fixtest = @"
+                public class MyClass
+                {
+                    public event System.EventHandler MyEvent;
+                    bool raiseEvents = true;
+
+                    public void Execute()
+                    {
+                        if (raiseEvents)
+                        {
+                            var handler = MyEvent;
+                            if (handler != null)
+                                handler(this, System.EventArgs.Empty);
+                        }
+                    }
+                }";
+
+            await VerifyCSharpFixAsync(source, fixtest, 1);
+        }
+
+        [Fact]
+        public async void OnlyOneFixIfExpressionBodied()
+        {
+            const string test = @"
+                public class MyClass
+                {
+                    public int Foo(int par1, int par2) => FuncCalc(par1,par2);
+
+                    public int FuncCalc(int p1, int p2)
+                    {
+                         return p1*p2;
+                    }
+                }";
+            await VerifyCSharpHasNumberOfCodeActions(test, 1);
+        }
+
+        [Fact]
+        public async void FixWhenInsideExpressionWithInvoke()
+        {
+            var code = @"
+public System.Func<string, bool> AllowInteraction { get; protected set; }
+protected bool AllowedInteraction()
+{
+    if (!AllowInteraction("""") || 1 == System.DateTime.Now.Second)
+    {
+        return false;
+    }
+    return true;
+}".WrapInCSharpClass();
+            var fix = @"
+public System.Func<string, bool> AllowInteraction { get; protected set; }
+protected bool AllowedInteraction()
+{
+    if (!AllowInteraction?.Invoke("""") ?? default(bool) || 1 == System.DateTime.Now.Second)
+    {
+        return false;
+    }
+    return true;
+}".WrapInCSharpClass();
+            await VerifyCSharpFixAsync(code, fix);
+        }
+
+        [Fact]
+        public async void FixWhenInsideABinaryExpressionWithPrecedenceWithInvoke()
+        {
+            var code = @"
+void Foo(Func<bool> f)
+{
+    var b = true;
+    if (b || f())
+    {
+    }
+}".WrapInCSharpClass();
+            var fix = @"
+void Foo(Func<bool> f)
+{
+    var b = true;
+    if (b || (f?.Invoke() ?? default(bool)))
+    {
+    }
+}".WrapInCSharpClass();
+            await VerifyCSharpFixAsync(code, fix);
+        }
+
+        [Fact]
+        public async void FixWhenInsideExpressionWithCheckForNull()
+        {
+            var code = @"
+public System.Func<string, bool> AllowInteraction { get; protected set; }
+protected bool AllowedInteraction()
+{
+    if (!AllowInteraction("""") || 1 == System.DateTime.Now.Second)
+    {
+        return false;
+    }
+    return true;
+}".WrapInCSharpClass();
+            var fix = @"
+public System.Func<string, bool> AllowInteraction { get; protected set; }
+protected bool AllowedInteraction()
+{
+    var handler = AllowInteraction;
+    if (handler != null)
+        if (!handler("""") || 1 == System.DateTime.Now.Second)
+        {
+            return false;
+        }
+    return true;
+}".WrapInCSharpClass();
+            await VerifyCSharpFixAsync(code, fix, 1);
+        }
+
+        [Fact]
+        public async void FixWhenInsideExpressionAndNameAlreadyExists()
+        {
+            var code = @"
+public System.Func<string, bool> AllowInteraction { get; protected set; }
+protected bool AllowedInteraction()
+{
+    var handler = 1;
+    if (!AllowInteraction("""") || 1 == System.DateTime.Now.Second)
+    {
+        return false;
+    }
+    return true;
+}".WrapInCSharpClass();
+            var fix = @"
+public System.Func<string, bool> AllowInteraction { get; protected set; }
+protected bool AllowedInteraction()
+{
+    var handler = 1;
+    var handler1 = AllowInteraction;
+    if (handler1 != null)
+        if (!handler1("""") || 1 == System.DateTime.Now.Second)
+        {
+            return false;
+        }
+    return true;
+}".WrapInCSharpClass();
+            await VerifyCSharpFixAsync(code, fix, 1);
+        }
+
+        [Fact]
+        public async void WhenEventIsFiredDirectlyShouldCopyItToVariable()
+        {
+            const string source = @"
+                public class MyClass
+                {
+                    public event System.EventHandler MyEvent;
+
+                    public void Execute()
+                    {
+                        MyEvent(this, System.EventArgs.Empty);
+                    }
+                }";
+
+            const string fixtest = @"
+                public class MyClass
+                {
+                    public event System.EventHandler MyEvent;
+
+                    public void Execute()
+                    {
+                        var handler = MyEvent;
+                        if (handler != null)
+                            handler(this, System.EventArgs.Empty);
+                    }
+                }";
+
+            await VerifyCSharpFixAsync(source, fixtest, 1);
+        }
+
+        [Fact]
+        public async void NotWarningIfEventIsCopiedToLocalVariableBeforeFire()
+        {
+            const string test = @"
+                public class MyClass
+                {
+                    public event System.EventHandler MyEvent;
+
+                    public void Execute()
+                    {
+                        var handler = MyEvent;
+                        if (handler != null)
+                            handler(this, System.EventArgs.Empty);
+                    }
+                }";
 
             await VerifyCSharpHasNoDiagnosticsAsync(test);
         }
